Finish detangling, it's fully broken, let's fix it
